+++ mobile/android/android-components/components/browser/engine-system/src/test/java/mozilla/components/browser/engine/system/SystemEngineViewTest.kt
@@ -350,6 +350,10 @@ class SystemEngineViewTest {
                 fail()
             }
 
+            override suspend fun onPreviewImageChange(uri: String, previewImageUrl: String) {
+                fail()
+            }
+
             override suspend fun getVisited(uris: List<String>): List<Boolean> {
                 fail()
                 return emptyList()