+++ mobile/android/android-components/components/browser/storage-sync/src/main/java/mozilla/components/browser/storage/sync/PlacesHistoryStorage.kt
@@ -77,7 +77,8 @@ open class PlacesHistoryStorage(
                     VisitObservation(
                         url = uri,
                         visitType = mozilla.appservices.places.VisitType.UPDATE_PLACE,
-                        title = observation.title
+                        title = observation.title,
+                        previewImageUrl = observation.previewImageUrl
                     )
                 )
             }