+++ mobile/android/android-components/components/concept/engine/src/test/java/mozilla/components/concept/engine/EngineSessionTest.kt
@@ -826,6 +826,7 @@ class EngineSessionTest {
 
         defaultObserver.onTitleChange("")
         defaultObserver.onLocationChange("")
+        defaultObserver.onPreviewImageChange("")
         defaultObserver.onLongPress(HitResult.UNKNOWN(""))
         defaultObserver.onExternalResource("", "")
         defaultObserver.onDesktopModeChange(true)